% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/haiku.R
\name{haiku}
\alias{haiku}
\title{Make me a haiku!}
\usage{
haiku()
}
\value{
a haiku
}
\description{
Prints haikus on a console. Remember though: Too much haikus and no fun makes
Jack a dull boy.
}
\examples{
#Just do
haiku()
}
